[PLAT-9074] Limit controls to right drawer
